Add explicit support for multi-step loading
There are cases where multiple FITs must be loaded to complete a boot,
e.g. where the first FIT contains devicetrees and the second contains
the OS.

At present this is not permitted, since every FIT must contain a
bootable kernel (or firmware).

Add a new 'load-only' property to indicate that a configuration can be
loaded but not booted. Add a table indicating how this should be
implemented. Describe how the option works in the 'usage' section.

Signed-off-by: Simon Glass <sjg@chromium.org>